+++ sphinx/domains/python/__init__.py
@@ -1082,7 +1082,7 @@ def setup(app: Sphinx) -> ExtensionMetadata:
         'python_trailing_comma_in_multi_line_signatures',
         True,
         'env',
-        bool,
+        types=frozenset({bool}),
     )
     app.add_config_value('python_display_short_literal_types', False, 'env')
     app.connect('object-description-transform', filter_meta_fields)